livekit-agent/pyproject.toml
[project]
name = "livekit-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[silero,turn-detector]~=1.2",
    "livekit-plugins-openai~=1.0",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv~=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
